adding jwt as a context
