+++ src/JwtContext.js
@@ -1,0 +1,4 @@
+import React from "react";
+
+const JwtContext = React.createContext();
+export default JwtContext;
